+++ spec/rubocop/cop/style/struct_inheritance_spec.rb
@@ -1,0 +1,36 @@
+# encoding: utf-8
+
+require 'spec_helper'
+
+describe RuboCop::Cop::Style::StructInheritance do
+  subject(:cop) { described_class.new }
+
+  it 'registers an offense when extending instance of Struct' do
+    inspect_source(cop,
+                   ['class Person < Struct.new(:first_name, :last_name)',
+                    'end'
+                   ])
+    expect(cop.offenses.size).to eq(1)
+  end
+
+  it 'registers an offense when extending instance of Struct' do
+    inspect_source(cop,
+                   ['class Person < Struct.new(:first_name, :last_name) do end',
+                    'end'
+                   ])
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'accepts plain class' do
+    inspect_source(cop,
+                   ['class Person',
+                    'end'
+                   ])
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'accepts assignment to Struct.new' do
+    inspect_source(cop, 'Person = Struct.new(:first_name, :last_name)')
+    expect(cop.offenses).to be_empty
+  end
+end